Fixed enemy bug
Fixed code, enemies can now get eliminated again when stepped on.
Separated boss jump check from common enemy jump check
