+++ Assets/Scripts/EnemiesAndObjects/BossLvlJumpCheck.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6e089eead45241145bc400904988cb4e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
