+++ Assets/Scripts/Player/EnemyHeadCheck.cs
@@ -9,10 +9,11 @@ public class EnemyHeadCheck : MonoBehaviour
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
-        if (collision.GetComponent<PlayerFeetCheck>())
+        if (collision.GetComponent<PlayerFeetCheck>() || collision.GetComponent<BossLvlJumpCheck>())
         {
             playerRb.velocity = new Vector2(playerRb.velocity.x, 0f);
             playerRb.AddForce(Vector2.up * 600f);
         }
+
     }
 }
